backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
azure-storage-blob==12.24.0
cryptography==44.0.0
apscheduler==3.10.4
python-dateutil==2.9.0
pytz==2024.2
